+++ wagtail/admin/tests/pages/test_explorer_view.py
@@ -5,10 +5,10 @@ from django.test import TestCase, override_settings
 from django.urls import reverse
 
 from wagtail import hooks
-from wagtail.admin.tests.pages.timestamps import local_datetime
 from wagtail.models import GroupPagePermission, Locale, Page
 from wagtail.test.testapp.models import SimplePage, SingleEventPage, StandardIndex
 from wagtail.test.utils import WagtailTestUtils
+from wagtail.test.utils.timestamps import local_datetime
 
 
 class TestPageExplorer(TestCase, WagtailTestUtils):
